Mise à jour pour DFCE 0.9.3
svn path=/java/trunk/; revision=1178

+++ dfcetest/src/test/java/fr/urssaf/image/tests/dfcetest/helpers/NcotiHelper.java
@@ -84,7 +84,7 @@ public class NcotiHelper {
       Category DATETIME_REF = storageAdministrationService.findOrCreateCategory(
             Categories.DATETIME.toString(), CategoryDataType.DATETIME);  
       Category DECIMAL_REF = storageAdministrationService.findOrCreateCategory(
-            Categories.DECIMAL.toString(), CategoryDataType.DECIMAL);
+            Categories.DOUBLE.toString(), CategoryDataType.DOUBLE);
       Category INTEGER_REF = storageAdministrationService.findOrCreateCategory(
             Categories.INTEGER.toString(), CategoryDataType.INTEGER); 
       Category BOOLEAN_REF = storageAdministrationService.findOrCreateCategory(
